// lib/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
      counters: {
        Row: {
          category_ids: string[]
          counter_number: number
          created_at: string
          id: string
          is_open: boolean
          queue_history: Json[]
        }
        Insert: {
          category_ids?: string[]
          counter_number?: number
          created_at?: string
          id?: string
          is_open?: boolean
          queue_history?: Json[]
        }
        Update: {
          category_ids?: string[]
          counter_number?: number
          created_at?: string
          id?: string
          is_open?: boolean
          queue_history?: Json[]
        }
        Relationships: []
      }
      queue_items: {
        Row: {
          category_id: string
          created_at: string
          id: string
          queue_number: string
          service_name: string
        }
        Insert: {
          category_id: string
          created_at?: string
          id?: string
          queue_number?: string
          service_name?: string
        }
        Update: {
          category_id?: string
          created_at?: string
          id?: string
          queue_number?: string
          service_name?: string
        }
        Relationships: [
          {
            foreignKeyName: "queueItems_categoryId_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "services"
            referencedColumns: ["id"]
          },
        ]
      }
      services: {
        Row: {
          children_ids: string[]
          created_at: string
          id: string
          level: number
          name: Json
          token_prefix: string
        }
        Insert: {
          children_ids?: string[]
          created_at?: string
          id?: string
          level?: number
          name?: Json
          token_prefix?: string
        }
        Update: {
          children_ids?: string[]
          created_at?: string
          id?: string
          level?: number
          name?: Json
          token_prefix?: string
        }
        Relationships: []
      }
    }
// ...
  }
}